Move utility methods to Bff.Util
